Trait tardigrade_worker::HandleRequest

pub trait HandleRequest<P: WorkerStoragePool>: WorkerInterface {
    // Required method
    fn handle_request<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        request: Request<Self>,
        connection: &'life1 mut P::Connection<'life2>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided methods
    fn initialize<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        connection_pool: &'life1 P
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn handle_cancellation<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        cancellation: Cancellation,
        connection: &'life1 mut P::Connection<'life2>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
}

Handler of requests of a certain type.

Required Methods

fn handle_request<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, request: Request<Self>, connection: &'life1 mut P::Connection<'life2> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Handles a request.

A request can be handled either synchronously (by calling Request::into_parts() and then the relevant ResponseSender method with connection), or asynchronously (by moving the ResponseSender into a task and creating a dedicated connection for sending the response).

Provided Methods

fn initialize<'life0, 'life1, 'async_trait>( &'life0 mut self, connection_pool: &'life1 P ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Initializes this handler. This method will be called and awaited before any other methods are called for the handler instance.

fn handle_cancellation<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, cancellation: Cancellation, connection: &'life1 mut P::Connection<'life2> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Handles request cancellation.