Trait tardigrade_worker::WorkerStorageConnection

pub trait WorkerStorageConnection: Send {
    type Error: Error + Send + Sync + 'static;

    // Required methods
    fn worker<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<WorkerRecord>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_or_create_worker<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<WorkerRecord, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_worker_cursor<'life0, 'async_trait>(
        &'life0 mut self,
        worker_id: u64,
        cursor: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn push_message<'life0, 'async_trait>(
        &'life0 mut self,
        channel_id: ChannelId,
        message: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn close_response_channel<'life0, 'async_trait>(
        &'life0 mut self,
        channel_id: ChannelId
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn release<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait;
}

Read-write connection to the worker storage.

This view is not required to be transactional, although it can be if the underlying storage supports this.

Required Associated Types

type Error: Error + Send + Sync + 'static

Errors that can occur when applying changes to the storage.

Required Methods

fn worker<'life0, 'life1, 'async_trait>( &'life0 mut self, name: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Option<WorkerRecord>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Gets a WorkerRecord by the (globally unique) worker name.

fn get_or_create_worker<'life0, 'life1, 'async_trait>( &'life0 mut self, name: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<WorkerRecord, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Gets or creates a WorkerRecord by the (globally unique) worker name.

fn update_worker_cursor<'life0, 'async_trait>( &'life0 mut self, worker_id: u64, cursor: u64 ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Updates the cursor value of a worker with the specified ID.

fn push_message<'life0, 'async_trait>( &'life0 mut self, channel_id: ChannelId, message: Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sends the message over a channel.

fn close_response_channel<'life0, 'async_trait>( &'life0 mut self, channel_id: ChannelId ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Closes a channel from the sender side. Returns true if the channel was actually closed and false otherwise (i.e., if the channel has other senders).

fn release<'async_trait>( self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait,

Releases the connection back to the pool.

If this connection is transactional, atomically commits changes to the storage.

If this connection is not transactional, the changes are expected to take effect immediately, and this method can be no-op.