Trait tardigrade_worker::WorkerStoragePool

pub trait WorkerStoragePool: Send + Sync + 'static {
    type Error: Error + Send + Sync + 'static;
    type Connection<'a>: WorkerStorageConnection<Error = Self::Error>
       where Self: 'a;

    // Required methods
    fn connect<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Self::Connection<'_>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stream_messages(
        &self,
        channel_id: ChannelId,
        start_idx: u64
    ) -> MessageStream<Self::Error>;
}

Connection pool for the worker storage.

Required Associated Types

type Error: Error + Send + Sync + 'static

Errors that can occur when applying changes to the storage.

type Connection<'a>: WorkerStorageConnection<Error = Self::Error> where Self: 'a

Read-write connection to the storage.

Required Methods

fn connect<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Self::Connection<'_>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Connects to the worker storage.

fn stream_messages( &self, channel_id: ChannelId, start_idx: u64 ) -> MessageStream<Self::Error>

Streams messages from a specific channel starting with the specified index.